+++ src/main/java/com/jnape/palatable/lambda/adt/Either.java
@@ -286,15 +286,13 @@ public abstract class Either<L, R> implements CoProduct2<L, R, Either<L, R>>, Mo
      * @param <L>      the left parameter type
      * @param <R>      the right parameter type
      * @return the supplier result as a right value, or leftFn's mapping result as a left value
+     * @deprecated in favor of {@link Try#trying(CheckedSupplier)}
      */
+    @Deprecated
     @SuppressWarnings("unchecked")
     public static <E extends Exception, L, R> Either<L, R> trying(CheckedSupplier<E, ? extends R> supplier,
                                                                   Function<? super E, ? extends L> leftFn) {
-        try {
-            return right(supplier.get());
-        } catch (Exception e) {
-            return left(leftFn.apply((E) e));
-        }
+        return Try.trying(supplier).toEither(leftFn).biMap(id(), id());
     }
 
     /**
@@ -305,7 +303,9 @@ public abstract class Either<L, R> implements CoProduct2<L, R, Either<L, R>>, Mo
      * @param <E>      the left parameter type (the most contravariant exception that supplier might throw)
      * @param <R>      the right parameter type
      * @return the supplier result as a right value, or a left value of the thrown exception
+     * @deprecated in favor of {@link Try#trying(CheckedSupplier)}
      */
+    @Deprecated
     public static <E extends Exception, R> Either<E, R> trying(CheckedSupplier<E, R> supplier) {
         return trying(supplier, id());
     }