+++ src/main/java/com/jnape/palatable/lambda/functions/specialized/checked/Runtime.java
@@ -3,7 +3,7 @@ package com.jnape.palatable.lambda.functions.specialized.checked;
 class Runtime {
 
     @SuppressWarnings("unchecked")
-    public static <T extends Throwable> RuntimeException throwChecked(Throwable ex) throws T {
-        throw (T) ex;
+    public static <T extends Throwable> RuntimeException throwChecked(Throwable t) throws T {
+        throw (T) t;
     }
 }